Client: refactor Transactions component to be used in account detail and transaction index page
